Added a ton of files and versions
Added a ton of files and versions of different components. I did so much work at once so that's why it was all committed right now. Yes/No check, ABCD Options checker, and Instructions are all components that have been added. This was all done on the first day back from school.